Move `StopAtSpec` to its own module
